clickup_to_bigquery/requirements.txt
google-cloud-bigquery==3.11.4
requests==2.31.0
python-dotenv==1.0.0 